requirements.txt
boto3>=1.35,<1.36
python-dotenv>=1.0,<2.0
streamlit>=1.41,<2.0
PyYAML>=6.0.2,<7.0
google-auth-oauthlib>=1.2.0,<2.0
google-auth>=2.27.0,<3.0
requests>=2.31.0,<3.0
